+++ src/anemoi/graphs/edges/attributes.py
@@ -9,103 +9,80 @@
 
 from __future__ import annotations
 
-from abc import ABC
-from abc import abstractmethod
+import logging
 
-import numpy as np
 import torch
-from torch_geometric.data import HeteroData
+from torch_geometric.nn.conv import MessagePassing
+from torch_geometric.typing import Adj
+from torch_geometric.typing import PairTensor
+from torch_geometric.typing import Size
 
-from anemoi.graphs.edges.directional import directional_edge_features
+from anemoi.graphs.edges.directional import compute_directions
 from anemoi.graphs.normalise import NormaliserMixin
 from anemoi.graphs.utils import haversine_distance
 
+LOGGER = logging.getLogger(__name__)
 
-class BaseEdgeAttribute(ABC, NormaliserMixin):
-    """Base class for edge attributes."""
+
+class BaseEdgeAttributeBuilder(MessagePassing, NormaliserMixin):
+    """Base class for edge attribute builders."""
 
     def __init__(self, norm: str | None = None) -> None:
+        super().__init__()
+        self._idx_lat = 0
+        self._idx_lon = 1
         self.norm = norm
 
-    @abstractmethod
-    def get_raw_values(self, graph: HeteroData, source_name: str, target_name: str, *args, **kwargs) -> np.ndarray: ...
+    def forward(self, x: PairTensor, edge_index: Adj, size: Size = None) -> torch.Tensor:
+        return self.propagate(edge_index, x=x, size=size)
 
-    def post_process(self, values: np.ndarray) -> torch.Tensor:
-        """Post-process the values."""
-        if values.ndim == 1:
-            values = values[:, np.newaxis]
-
-        normed_values = self.normalise(values)
-
-        return torch.tensor(normed_values, dtype=torch.float32)
-
-    def compute(self, graph: HeteroData, edges_name: tuple[str, str, str], *args, **kwargs) -> torch.Tensor:
-        """Compute the edge attributes."""
-        source_name, _, target_name = edges_name
-        assert (
-            source_name in graph.node_types
-        ), f"Node \"{source_name}\" not found in graph. Optional nodes are {', '.join(graph.node_types)}."
-        assert (
-            target_name in graph.node_types
-        ), f"Node \"{target_name}\" not found in graph. Optional nodes are {', '.join(graph.node_types)}."
-
-        values = self.get_raw_values(graph, source_name, target_name, *args, **kwargs)
-        return self.post_process(values)
-
-
-class EdgeDirection(BaseEdgeAttribute):
-    """Edge direction feature.
-
-    This class calculates the direction of an edge using either:
-    1. Rotated features: The target nodes are rotated to the north pole to compute the edge direction.
-    2. Non-rotated features: The direction is computed as the difference in latitude and longitude between the source
-    and target nodes.
-
-    The resulting direction is represented as a unit vector starting at (0, 0), with X and Y components.
-
-    Attributes
-    ----------
-    norm : Optional[str]
-        Normalisation method. Options: None, "l1", "l2", "unit-max", "unit-range", "unit-std".
-    luse_rotated_features : bool
-        Whether to use rotated features.
-
-    Methods
-    -------
-    compute(graph, source_name, target_name)
-        Compute direction of all edges.
-    """
-
-    def __init__(self, norm: str | None = None, luse_rotated_features: bool = True) -> None:
-        super().__init__(norm)
-        self.luse_rotated_features = luse_rotated_features
-
-    def get_raw_values(self, graph: HeteroData, source_name: str, target_name: str) -> np.ndarray:
-        """Compute directional features for edges.
+    def compute(self, x_i: torch.Tensor, x_j: torch.Tensor) -> torch.Tensor:
+        """Compute edge features.
 
         Parameters
         ----------
-        graph : HeteroData
-            The graph.
-        source_name : str
-            The name of the source nodes.
-        target_name : str
-            The name of the target nodes.
+        x_i : torch.Tensor
+            Coordinates of the source nodes.
+        x_j : torch.Tensor
+            Coordinates of the target nodes.
 
         Returns
         -------
-        np.ndarray
-            The directional features.
+        torch.Tensor
+            Edge features.
         """
-        edge_index = graph[(source_name, "to", target_name)].edge_index
-        source_coords = graph[source_name].x.numpy()[edge_index[0]].T
-        target_coords = graph[target_name].x.numpy()[edge_index[1]].T
-        edge_dirs = directional_edge_features(source_coords, target_coords, self.luse_rotated_features).T
+        raise NotImplementedError("Method `compute` must be implemented.")
+
+    def message(self, x_i: torch.Tensor, x_j: torch.Tensor) -> torch.Tensor:
+        edge_features = self.compute(x_i, x_j)
+
+        if edge_features.ndim == 1:
+            edge_features = edge_features.unsqueeze(-1)
+
+        return self.normalise(edge_features)
+
+    def aggregate(self, edge_features: torch.Tensor) -> torch.Tensor:
+        return edge_features
+
+
+class EdgeLength(BaseEdgeAttributeBuilder):
+    """Computes edge length for bipartite graphs."""
+
+    def compute(self, x_i: torch.Tensor, x_j: torch.Tensor) -> torch.Tensor:
+        edge_length = haversine_distance(x_i, x_j)
+        return edge_length
+
+
+class EdgeDirection(BaseEdgeAttributeBuilder):
+    """Computes edge direction for bipartite graphs."""
+
+    def compute(self, x_i: torch.Tensor, x_j: torch.Tensor) -> torch.Tensor:
+        edge_dirs = compute_directions(x_i, x_j)
         return edge_dirs
 
 
-class EdgeLength(BaseEdgeAttribute):
-    """Edge length feature.
+class Azimuth(BaseEdgeAttributeBuilder):
+    """Compute the azimuth of the edge.
 
     Attributes
     ----------
@@ -118,40 +95,22 @@ class EdgeLength(BaseEdgeAttribute):
     -------
     compute(graph, source_name, target_name)
         Compute edge lengths attributes.
+
+    References
+    ----------
+    - https://www.movable-type.co.uk/scripts/latlong.html
     """
 
-    def __init__(self, norm: str | None = None, invert: bool = False) -> None:
-        super().__init__(norm)
-        self.invert = invert
+    def compute(self, x_i: torch.Tensor, x_j: torch.Tensor) -> torch.Tensor:
+        # Forward bearing. x_i, x_j must bez radians.
+        a11 = torch.cos(x_i[:, self._idx_lat]) * torch.sin(x_j[:, self._idx_lat])
+        a12 = (
+            torch.sin(x_i[:, self._idx_lat])
+            * torch.cos(x_j[:, self._idx_lat])
+            * torch.cos(x_j[..., self._idx_lon] - x_i[..., self._idx_lon])
+        )
+        a1 = a11 - a12
+        a2 = torch.sin(x_j[..., self._idx_lon] - x_i[..., self._idx_lon]) * torch.cos(x_j[:, self._idx_lat])
+        edge_dirs = torch.atan2(a2, a1)
 
-    def get_raw_values(self, graph: HeteroData, source_name: str, target_name: str) -> np.ndarray:
-        """Compute haversine distance (in kilometers) between nodes connected by edges.
-
-        Parameters
-        ----------
-        graph : HeteroData
-            The graph.
-        source_name : str
-            The name of the source nodes.
-        target_name : str
-            The name of the target nodes.
-
-        Returns
-        -------
-        np.ndarray
-            The edge lengths.
-        """
-        edge_index = graph[(source_name, "to", target_name)].edge_index
-        source_coords = graph[source_name].x.numpy()[edge_index[0]]
-        target_coords = graph[target_name].x.numpy()[edge_index[1]]
-        edge_lengths = haversine_distance(source_coords, target_coords)
-        return edge_lengths
-
-    def post_process(self, values: np.ndarray) -> torch.Tensor:
-        """Post-process edge lengths."""
-        values = super().post_process(values)
-
-        if self.invert:
-            values = 1 - values
-
-        return values
+        return edge_dirs