+++ src/anemoi/graphs/nodes/builders/base.py
@@ -9,6 +9,8 @@
 
 from __future__ import annotations
 
+import logging
+import time
 from abc import ABC
 from abc import abstractmethod
 
@@ -17,6 +19,10 @@ import torch
 from anemoi.utils.config import DotDict
 from hydra.utils import instantiate
 from torch_geometric.data import HeteroData
+
+from anemoi.graphs.utils import get_grid_reference_distance
+
+LOGGER = logging.getLogger(__name__)
 
 
 class BaseNodeBuilder(ABC):
@@ -53,6 +59,7 @@ class BaseNodeBuilder(ABC):
         """
         graph[self.name].x = self.get_coordinates()
         graph[self.name].node_type = type(self).__name__
+        graph[self.name]["_grid_reference_distance"] = get_grid_reference_distance(graph[self.name].x.cpu())
         return graph
 
     def register_attributes(self, graph: HeteroData, config: DotDict | None = None) -> HeteroData:
@@ -115,11 +122,17 @@ class BaseNodeBuilder(ABC):
         HeteroData
             The graph with new nodes included.
         """
+        t0 = time.time()
         graph = self.register_nodes(graph)
+        t1 = time.time()
+        LOGGER.debug("Time to register node coordinates (%s): %.2f s", self.__class__.__name__, t1 - t0)
 
         if attrs_config is None:
             return graph
 
+        t0 = time.time()
         graph = self.register_attributes(graph, attrs_config)
+        t1 = time.time()
+        LOGGER.debug("Time to register node coordinates (%s): %.2f s", self.__class__.__name__, t1 - t0)
 
         return graph